Fix for Quality/Signal level output variations
As mentioned in https://github.com/KanoComputing/kano-toolset/issues/198, the format of Quality and Signal level changes when iwlist cannot get a range for some Access Points (or perhaps it's related to the wireless driver). At any rate, I've observed a 2Wire/AT&T U-verse router showing:

`Quality:40  Signal level:0  Noise level:0`

The original parser only accounts for Quality/Signal level in this format:

`Quality=72/100  Signal level=65/100`

This change accommodates the possibility of for a colon in addition to an equals sign, and also can handle an integer value for Quality as opposed to a fraction.